+++ src/test/java/uk/gov/hmcts/dts/fact/entity/EmailTest.java
@@ -1,34 +1,16 @@
 package uk.gov.hmcts.dts.fact.entity;
 
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
-import org.springframework.context.i18n.LocaleContextHolder;
-
-import java.util.Locale;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
-class EmailTest {
-
-    static uk.gov.hmcts.dts.fact.entity.Email entity;
-
-    @BeforeAll
-    static void setUp() {
-        entity = new uk.gov.hmcts.dts.fact.entity.Email();
-        entity.setAddress("test@test.com");
-        entity.setDescription("An email address");
-        entity.setDescriptionCy("An email address but in Welsh");
-        entity.setExplanation("You email it.");
-        entity.setExplanationCy("You email it in Welsh.");
-    }
+public class EmailTest {
 
     @Test
     void testEmailTypeConstructor() {
         EmailType emailType = new EmailType(1, "desc", "desc cy");
-        uk.gov.hmcts.dts.fact.entity.Email email = new uk.gov.hmcts.dts.fact.entity.Email(
+        Email email = new Email(
             "address", "expl", "expl cy", emailType);
         assertEquals(email.getAdminEmailType(), emailType);
         assertEquals(email.getExplanation(), "expl");
@@ -37,20 +19,4 @@ class EmailTest {
         assertEquals(email.getDescription(), "");
         assertNull(email.getDescriptionCy());
     }
-
-    @ParameterizedTest
-    @ValueSource(booleans = {false, true})
-    void testCreation(boolean welsh) {
-        if (welsh) {
-            Locale locale = new Locale("cy");
-            LocaleContextHolder.setLocale(locale);
-        }
-        uk.gov.hmcts.dts.fact.model.Email email = new uk.gov.hmcts.dts.fact.model.Email(entity);
-        assertEquals(entity.getAddress(), email.getAddress());
-        assertEquals(welsh ? entity.getExplanationCy() : entity.getExplanation(), email.getExplanation());
-        assertEquals(welsh ? entity.getDescriptionCy() : entity.getDescription(), email.getDescription());
-
-        LocaleContextHolder.resetLocaleContext();
-    }
-
 }
